#26844 Renaming api-data-model com.dotcms.api.client package.
