+++ tools/dotcms-cli/api-data-model/src/main/java/com/dotcms/api/DefaultAuthenticationContextImpl.java
@@ -1,9 +1,9 @@
 package com.dotcms.api;
 
 
-import com.dotcms.api.client.AuthenticationParam;
-import com.dotcms.api.client.RestClientFactory;
-import com.dotcms.api.client.ServiceManager;
+import com.dotcms.api.client.model.AuthenticationParam;
+import com.dotcms.api.client.model.RestClientFactory;
+import com.dotcms.api.client.model.ServiceManager;
 import com.dotcms.model.ResponseEntityView;
 import com.dotcms.model.annotation.SecuredPassword;
 import com.dotcms.model.authentication.APITokenRequest;
