+++ tools/dotcms-cli/api-data-model/src/test/java/com/dotcms/api/client/model/ServiceManagerTest.java
@@ -1,5 +1,6 @@
-package com.dotcms.api.client;
+package com.dotcms.api.client.model;
 
+import com.dotcms.api.client.model.ServiceManager;
 import com.dotcms.model.config.CredentialsBean;
 import com.dotcms.model.config.ServiceBean;
 import io.quarkus.test.junit.QuarkusTest;
