+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/files/PullBase.java
@@ -1,7 +1,7 @@
 package com.dotcms.api.client.files;
 
 import com.dotcms.api.LanguageAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.client.files.traversal.LocalTraversalService;
 import com.dotcms.api.traversal.TreeNode;
 import com.dotcms.api.traversal.TreeNodeInfo;
