+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/files/traversal/data/Pusher.java
@@ -9,7 +9,7 @@ import com.dotcms.api.AssetAPI;
 import com.dotcms.api.FolderAPI;
 import com.dotcms.api.LanguageAPI;
 import com.dotcms.api.SiteAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.model.ResponseEntityView;
 import com.dotcms.model.asset.AssetView;
 import com.dotcms.model.asset.ByPathRequest;
