+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/pull/language/LanguageFetcher.java
@@ -1,7 +1,7 @@
 package com.dotcms.api.client.pull.language;
 
 import com.dotcms.api.LanguageAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.client.pull.ContentFetcher;
 import com.dotcms.model.language.Language;
 import java.io.Serializable;
