+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/pull/site/SiteFetcher.java
@@ -1,7 +1,7 @@
 package com.dotcms.api.client.pull.site;
 
 import com.dotcms.api.SiteAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.client.pull.ContentFetcher;
 import com.dotcms.model.ResponseEntityView;
 import com.dotcms.model.site.GetSiteByNameRequest;
