+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/api/client/push/contenttype/ContentTypePushHandler.java
@@ -1,7 +1,7 @@
 package com.dotcms.api.client.push.contenttype;
 
 import com.dotcms.api.ContentTypeAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.client.push.PushHandler;
 import com.dotcms.contenttype.model.type.ContentType;
 import com.dotcms.model.contenttype.AbstractSaveContentTypeRequest.Builder;
