+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/cli/command/EntryCommand.java
@@ -1,6 +1,6 @@
 package com.dotcms.cli.command;
 
-import com.dotcms.api.client.AuthenticationParam;
+import com.dotcms.api.client.model.AuthenticationParam;
 import com.dotcms.cli.command.contenttype.ContentTypeCommand;
 import com.dotcms.cli.command.files.FilesCommand;
 import com.dotcms.cli.command.language.LanguageCommand;
