+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/cli/command/InstanceCommand.java
@@ -1,7 +1,7 @@
 package com.dotcms.cli.command;
 
-import com.dotcms.api.client.DotCmsClientConfig;
-import com.dotcms.api.client.ServiceManager;
+import com.dotcms.api.client.model.DotCmsClientConfig;
+import com.dotcms.api.client.model.ServiceManager;
 import com.dotcms.cli.common.HelpOptionMixin;
 import com.dotcms.cli.common.OutputOptionMixin;
 import com.dotcms.model.annotation.SecuredPassword;
