+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/cli/command/language/AbstractLanguageCommand.java
@@ -1,7 +1,7 @@
 package com.dotcms.cli.command.language;
 
 import com.dotcms.api.LanguageAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.cli.common.HelpOptionMixin;
 import com.dotcms.cli.common.OutputOptionMixin;
 import com.dotcms.model.language.Language;
