+++ tools/dotcms-cli/cli/src/main/java/com/dotcms/cli/command/language/LanguageFind.java
@@ -1,7 +1,7 @@
 package com.dotcms.cli.command.language;
 
 import com.dotcms.api.LanguageAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.cli.command.DotCommand;
 import com.dotcms.cli.common.OutputOptionMixin;
 import com.dotcms.model.language.Language;
