+++ tools/dotcms-cli/cli/src/test/java/com/dotcms/api/client/files/PullServiceIT.java
@@ -2,7 +2,7 @@ package com.dotcms.api.client.files;
 
 import com.dotcms.DotCMSITProfile;
 import com.dotcms.api.AuthenticationContext;
-import com.dotcms.api.client.ServiceManager;
+import com.dotcms.api.client.model.ServiceManager;
 import com.dotcms.api.client.files.traversal.RemoteTraversalService;
 import com.dotcms.cli.common.FilesTestHelper;
 import com.dotcms.cli.common.OutputOptionMixin;
