+++ tools/dotcms-cli/cli/src/test/java/com/dotcms/cli/command/language/LanguageCommandIT.java
@@ -3,7 +3,7 @@ package com.dotcms.cli.command.language;
 import com.dotcms.DotCMSITProfile;
 import com.dotcms.api.AuthenticationContext;
 import com.dotcms.api.LanguageAPI;
-import com.dotcms.api.client.RestClientFactory;
+import com.dotcms.api.client.model.RestClientFactory;
 import com.dotcms.api.provider.ClientObjectMapper;
 import com.dotcms.api.provider.YAMLMapperSupplier;
 import com.dotcms.cli.command.CommandTest;
